requirements.txt
python-telegram-bot==13.15
aiohttp
requests
edge-tts
beautifulsoup4
flask
